LPS-67296 Move sql generation code to the insert sql builder, cleansing DBProvider interface
