+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/exporter/InsertSQLBuilder.java
@@ -16,7 +16,12 @@ package com.liferay.portal.tools.data.partitioning.sql.builder.exporter;
 
 import com.liferay.portal.tools.data.partitioning.sql.builder.internal.exporter.SQLBuilder;
 
+import java.sql.Date;
 import java.sql.ResultSetMetaData;
+import java.sql.Timestamp;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 
 /**
  * @author Manuel de la Peña
@@ -51,6 +56,47 @@ public class InsertSQLBuilder implements SQLBuilder {
 		sb.append(")");
 
 		return sb.toString() + ";\n";
+	}
+
+	@Override
+	public String getDateTimeFormat() {
+		return "YYYY-MM-DD HH:MM:SS";
+	}
+
+	@Override
+	public String serializeTableField(Object field) {
+		StringBuilder sb = new StringBuilder();
+
+		if (field == null) {
+			sb.append("null");
+		}
+		else if ((field instanceof Date) || (field instanceof Timestamp)) {
+			sb.append("'");
+			sb.append(formatDateTime(field));
+			sb.append("'");
+		}
+		else if (field instanceof String) {
+			String value = (String)field;
+
+			value = value.replace("'", "''");
+
+			sb.append("'");
+			sb.append(value);
+			sb.append("'");
+		}
+		else {
+			sb.append("'");
+			sb.append(field);
+			sb.append("'");
+		}
+
+		return sb.toString();
+	}
+
+	protected String formatDateTime(Object date) {
+		DateFormat dateFormat = new SimpleDateFormat(getDateTimeFormat());
+
+		return dateFormat.format(date);
 	}
 
 }
