+++ modules/util/portal-tools-data-partitioning-sql-builder/src/main/java/com/liferay/portal/tools/data/partitioning/sql/builder/internal/exporter/SQLBuilder.java
@@ -24,4 +24,8 @@ public interface SQLBuilder {
 	public String build(
 		ResultSetMetaData resultSetMetaData, String tableName, String[] fields);
 
+	public String getDateTimeFormat();
+
+	public String serializeTableField(Object field);
+
 }
